Переделан в соответствии с общей версией
