+++ ST/Stud.Constantin/Project/src/components/ITEMS.js
@@ -1,6 +1,10 @@
-export default class BasketItem{
+import Item from './Item.js';
+
+export class CatalogItem extends Item{}
+
+export class BasketItem extends Item{
     constructor(item){
-        this.item=item;
+        super(item);
     }
     render(){
         return  `
@@ -27,5 +31,4 @@ export default class BasketItem{
         </div>
         `
     }
-
 }
